Add files via upload
Hello, I wrote this python code to create the Connect 4 Game for the Pirple: python is easy course project
and I have problems with exiting the while loop. The program should stop after one player has won the game, but somehow it continues. Something must be going wrong with the return statement of the check_for_win function... This function should return End  = 1 if there is a winner and the: while End == 0 loop should therefore be interrupted.
Can someone pleas help?
Thanks a lot.